getting somewhere, but this is confusing.
Changing textures in the point_cloud_renderer GLSL to be either index 0
or 1 (pointCloudTexture0 or pointCloudTexture1, each corresponding to a
camera) to the opposite of the skip block in the loop on line 611 will
show a perfect shaded image from the non-skipped camera.

This suggests that both cameras are sending texture and color data.
However...

Trying to do a linear combination from non-skipped color data is WEIRD
though, and not as I'd expect.