fix logout
The reason this is so labyrinthine is because of the users and billing
components. They bind dynamic behaviors that are defined in
storyteller-web (one of them being clearing the redis session cache).

Actually, the users component _didn't_, because it was defined in the
billing component. So I hoisted it out into a new component library.

In the fullness of time I imagine all of this will be consolidated back
under storyteller-web so we don't have to worry about binding actions at
a distance.

Anyway, this is tested and works locally.