critical drag and drop bug fix (#431)
fixed the annoying bug (happens on Windows only?) where u cannot drag
and drop it and it shows the stop cursor, when u let go of left click,
its stuck to the cursor and u have to click again for it to be added to
the scene. and dragging a character and dropping on the panel it self
opens the upload dialog. happens with objects and dragging animation
onto the timeline too.

solution: css fix by disabling text and image highlighting/dragging +
styling updates

bug video reference:

https://cdn.discordapp.com/attachments/1221845845540995164/1243819302910558209/2024-05-25_13-51-14.mp4?ex=6654d6be&is=6653853e&hm=642ba65d8113300a6b2f254f4e6e483005bd1a1ccd7392ede9896c5d38e9e550&