Fix / Flush Background Loaded Media on Logout (#560)
This is fix to =: When you log off and get on another account, it
continues to load the audio/movies from the previous account that you
just logged out from. It seems to be limited only to audio and movie
list; scenes, characters, animations and expressions seem fine. This bug
is currently present on production.

Co-authored-by: wil <wil@storyteller.ai>