Feature/expressions fix to (#579)
QOL Panel Changes
The style name is shown when rendering a movie, missing title due to
issues with the API.
The progress bar moves to 0 to 40 then 100, but this is still fake.

Added Animation Type to the Characters so they would not show
expressions unless they had it.
ROKO and Storygirl need their metadata updated for this. TEAM TESTING
ONLY Will have this metadata.

Changed Expression to Facial Expression.

Changed Lip Sync to Vocals / Speech

Changed Copy to reflect Timeline Track Changes.
I removed the Mannequins since they do not work well with the scenes.